refactor(ui): split monolith files (types.ts, settings, queues)
Split three large UI files into focused modules:

types.ts (2,038 → 14 domain modules with barrel re-export):
  agent-types, app-types, auth-types, capability-types, command-types,
  common-types, durable-types, event-types, image-types, llm-types,
  mcp-types, message-types, session-types, skill-types

settings/providers/page.tsx (966 → 4 files):
  model-row, provider-card, provider-dialogs + slim page coordinator

durable/queues/page.tsx (915 → 5 files):
  dlq-row, task-row, queue-stats-card, enqueue-dialog + slim page

All existing imports continue to work via barrel re-exports.

Fixes EVE-120
