+++ apps/ui/src/app/(main)/settings/providers/page.tsx
@@ -2,688 +2,28 @@
 
 import { useState } from "react";
 import { Button } from "@/components/ui/button";
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
-import { Badge } from "@/components/ui/badge";
+import { Card, CardContent } from "@/components/ui/card";
 import { Skeleton } from "@/components/ui/skeleton";
-import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Select, SelectContent, SelectItem, SelectTrigger } from "@/components/ui/select";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
 import { useQueryClient } from "@tanstack/react-query";
 import {
   useLlmProviders,
   useLlmModels,
-  useCreateLlmProvider,
-  useUpdateLlmProvider,
   useDeleteLlmProvider,
   useSyncProviderModels,
-  useCreateLlmModel,
   useDeleteLlmModel,
 } from "@/hooks/use-llm-providers";
 import { updateLlmModel } from "@/lib/api/llm-providers";
 import { useOrganization, useUpdateOrganization } from "@/hooks/use-organizations";
 import { queryKeys } from "@/lib/query-keys";
-import {
-  Plus,
-  Server,
-  Trash2,
-  Key,
-  Cpu,
-  Brain,
-  Wrench,
-  Image,
-  DollarSign,
-  Layers,
-  Info,
-  ChevronDown,
-  ChevronUp,
-  RefreshCw,
-  Download,
-  X,
-} from "lucide-react";
-import { ProviderIcon, getProviderLabel } from "@/components/providers/provider-icon";
-import { formatTokens } from "@/lib/formatting";
-import type {
-  LlmProvider,
-  LlmModelWithProvider,
-  LlmProviderType,
-  CreateLlmProviderRequest,
-  CreateLlmModelRequest,
-} from "@/lib/api/types";
+import { Plus, Server, Cpu } from "lucide-react";
+import { ProviderIcon } from "@/components/providers/provider-icon";
+import type { LlmProvider } from "@/lib/api/types";
 
-const PROVIDER_TYPES: { value: LlmProviderType; label: string }[] = [
-  { value: "openai", label: "OpenAI (Responses API)" },
-  { value: "openai_completions", label: "OpenAI (Completions API)" },
-  { value: "anthropic", label: "Anthropic" },
-  { value: "gemini", label: "Google Gemini" },
-];
-
-// Get API key placeholder based on provider type
-function getApiKeyPlaceholder(providerType: LlmProviderType): string {
-  switch (providerType) {
-    case "openai":
-    case "openai_completions":
-      return "sk-...";
-    case "anthropic":
-      return "sk-ant-api03-...";
-    default:
-      return "your-api-key";
-  }
-}
-
-function ProviderCard({
-  provider,
-  onDelete,
-  onSetApiKey,
-  onSyncModels,
-  isSyncing,
-}: {
-  provider: LlmProvider;
-  onDelete: (id: string) => void;
-  onSetApiKey: (provider: LlmProvider) => void;
-  onSyncModels: (id: string) => void;
-  isSyncing: boolean;
-}) {
-  // Only show sync button for providers without custom base URL (standard providers)
-  const canSync = !provider.base_url && provider.api_key_set;
-
-  return (
-    <Card>
-      <CardHeader className="flex flex-row items-start justify-between space-y-0">
-        <div className="flex items-center gap-3">
-          <ProviderIcon providerType={provider.provider_type} size="md" />
-          <div>
-            <CardTitle className="text-lg">{provider.name}</CardTitle>
-            <CardDescription className="text-sm">
-              {getProviderLabel(provider.provider_type)}
-            </CardDescription>
-          </div>
-        </div>
-        <Badge
-          variant="outline"
-          className={
-            provider.status === "active"
-              ? "bg-green-100 text-green-800"
-              : "bg-gray-100 text-gray-800"
-          }
-        >
-          {provider.status}
-        </Badge>
-      </CardHeader>
-      <CardContent>
-        <div className="space-y-2 text-sm">
-          {provider.base_url && (
-            <p className="text-muted-foreground truncate">URL: {provider.base_url}</p>
-          )}
-          <div className="flex items-center gap-2">
-            <Key className="h-4 w-4 text-muted-foreground" />
-            <span className="text-muted-foreground">
-              API Key: {provider.api_key_set ? "Configured" : "Not set"}
-            </span>
-          </div>
-        </div>
-        <div className="flex items-center justify-end gap-2 mt-4">
-          {canSync && (
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => onSyncModels(provider.id)}
-              disabled={isSyncing}
-              title="Discover available models from provider API"
-            >
-              <RefreshCw className={`h-4 w-4 mr-1 ${isSyncing ? "animate-spin" : ""}`} />
-              {isSyncing ? "Syncing..." : "Sync Models"}
-            </Button>
-          )}
-          <Button variant="outline" size="sm" onClick={() => onSetApiKey(provider)}>
-            <Key className="h-4 w-4 mr-1" />
-            {provider.api_key_set ? "Update Key" : "Set Key"}
-          </Button>
-          <Button
-            variant="ghost"
-            size="sm"
-            className="text-destructive"
-            onClick={() => onDelete(provider.id)}
-          >
-            <Trash2 className="h-4 w-4" />
-          </Button>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
-function formatCost(cost: number): string {
-  if (cost >= 100) {
-    return `$${cost.toFixed(0)}`;
-  }
-  if (cost >= 1) {
-    return `$${cost.toFixed(2)}`;
-  }
-  return `$${cost.toFixed(3)}`;
-}
-
-function ModelRow({
-  model,
-  onDelete,
-  onToggleInstalled,
-  isTogglingInstalled,
-}: {
-  model: LlmModelWithProvider;
-  onDelete: (id: string) => void;
-  onToggleInstalled: (id: string, installed: boolean) => void;
-  isTogglingInstalled: boolean;
-}) {
-  const [expanded, setExpanded] = useState(false);
-  const profile = model.profile;
-
-  return (
-    <div className="border rounded-lg overflow-hidden">
-      <div className="flex items-center justify-between p-3">
-        <div className="flex items-center gap-3">
-          <ProviderIcon
-            providerType={model.provider_type}
-            size="sm"
-            showBackground={false}
-            className="text-muted-foreground"
-          />
-          <div>
-            <div className="font-medium flex items-center gap-2">
-              {model.display_name}
-              {model.installed && (
-                <Badge
-                  variant="outline"
-                  className="text-xs bg-green-50 text-green-700 border-green-200"
-                >
-                  Installed
-                </Badge>
-              )}
-              {profile && (
-                <Badge
-                  variant="outline"
-                  className="text-xs bg-blue-50 text-blue-700 border-blue-200"
-                >
-                  {profile.family}
-                </Badge>
-              )}
-            </div>
-            <div className="text-sm text-muted-foreground">
-              {model.model_id} - {model.provider_name}
-            </div>
-          </div>
-        </div>
-        <div className="flex items-center gap-2">
-          {/* Profile capability badges */}
-          {profile && (
-            <div className="flex gap-1">
-              {profile.reasoning && (
-                <Badge variant="secondary" className="text-xs" title="Reasoning">
-                  <Brain className="h-3 w-3 mr-1" />
-                  Reasoning
-                </Badge>
-              )}
-              {profile.tool_call && (
-                <Badge variant="secondary" className="text-xs" title="Tool Calling">
-                  <Wrench className="h-3 w-3 mr-1" />
-                  Tools
-                </Badge>
-              )}
-              {profile.attachment && (
-                <Badge variant="secondary" className="text-xs" title="Attachments">
-                  <Image className="h-3 w-3 mr-1" />
-                  Vision
-                </Badge>
-              )}
-            </div>
-          )}
-          {/* Legacy capabilities (only show if no profile) */}
-          {!profile && model.capabilities.length > 0 && (
-            <div className="flex gap-1">
-              {model.capabilities.slice(0, 2).map((cap) => (
-                <Badge key={cap} variant="secondary" className="text-xs">
-                  {cap}
-                </Badge>
-              ))}
-              {model.capabilities.length > 2 && (
-                <Badge variant="secondary" className="text-xs">
-                  +{model.capabilities.length - 2}
-                </Badge>
-              )}
-            </div>
-          )}
-          <Badge
-            variant="outline"
-            className={
-              model.status === "active"
-                ? "bg-green-100 text-green-800"
-                : "bg-gray-100 text-gray-800"
-            }
-          >
-            {model.status}
-          </Badge>
-          {/* Install / Uninstall toggle */}
-          <Button
-            variant={model.installed ? "outline" : "default"}
-            size="sm"
-            onClick={() => onToggleInstalled(model.id, !model.installed)}
-            disabled={isTogglingInstalled}
-            title={
-              model.installed
-                ? "Uninstall model (remove from UI pickers)"
-                : "Install model (make available in UI pickers)"
-            }
-          >
-            {model.installed ? (
-              <>
-                <X className="h-4 w-4 mr-1" />
-                Uninstall
-              </>
-            ) : (
-              <>
-                <Download className="h-4 w-4 mr-1" />
-                Install
-              </>
-            )}
-          </Button>
-          {profile && (
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={() => setExpanded(!expanded)}
-              title={expanded ? "Collapse" : "Expand profile details"}
-            >
-              {expanded ? <ChevronUp className="h-4 w-4" /> : <ChevronDown className="h-4 w-4" />}
-            </Button>
-          )}
-          <Button
-            variant="ghost"
-            size="sm"
-            className="text-destructive"
-            onClick={() => onDelete(model.id)}
-          >
-            <Trash2 className="h-4 w-4" />
-          </Button>
-        </div>
-      </div>
-
-      {/* Expanded profile details */}
-      {expanded && profile && (
-        <div className="border-t bg-muted/30 p-4">
-          <div className="grid grid-cols-2 md:grid-cols-4 gap-4 text-sm mb-3">
-            {/* Limits */}
-            {profile.limits && (
-              <div>
-                <div className="font-medium text-muted-foreground mb-1 flex items-center gap-1">
-                  <Layers className="h-3.5 w-3.5" />
-                  Token Limits
-                </div>
-                <div>Context: {formatTokens(profile.limits.context)}</div>
-                {profile.limits.input && <div>Input: {formatTokens(profile.limits.input)}</div>}
-                <div>Output: {formatTokens(profile.limits.output)}</div>
-              </div>
-            )}
-
-            {/* Cost */}
-            {profile.cost && (
-              <div>
-                <div className="font-medium text-muted-foreground mb-1 flex items-center gap-1">
-                  <DollarSign className="h-3.5 w-3.5" />
-                  Cost / 1M Tokens
-                </div>
-                <div>Input: {formatCost(profile.cost.input)}</div>
-                <div>Output: {formatCost(profile.cost.output)}</div>
-                {profile.cost.cache_read && <div>Cache: {formatCost(profile.cost.cache_read)}</div>}
-              </div>
-            )}
-
-            {/* Capabilities */}
-            <div>
-              <div className="font-medium text-muted-foreground mb-1 flex items-center gap-1">
-                <Wrench className="h-3.5 w-3.5" />
-                Capabilities
-              </div>
-              <div className="space-y-0.5">
-                <div className={profile.tool_call ? "text-green-700" : "text-muted-foreground"}>
-                  {profile.tool_call ? "✓" : "✗"} Tool Calling
-                </div>
-                <div
-                  className={profile.structured_output ? "text-green-700" : "text-muted-foreground"}
-                >
-                  {profile.structured_output ? "✓" : "✗"} Structured Output
-                </div>
-                <div className={profile.reasoning ? "text-green-700" : "text-muted-foreground"}>
-                  {profile.reasoning ? "✓" : "✗"} Reasoning
-                </div>
-                <div className={profile.attachment ? "text-green-700" : "text-muted-foreground"}>
-                  {profile.attachment ? "✓" : "✗"} Attachments
-                </div>
-              </div>
-            </div>
-
-            {/* Info */}
-            <div>
-              <div className="font-medium text-muted-foreground mb-1 flex items-center gap-1">
-                <Info className="h-3.5 w-3.5" />
-                Model Info
-              </div>
-              {profile.knowledge && <div>Knowledge: {profile.knowledge}</div>}
-              {profile.release_date && <div>Released: {profile.release_date}</div>}
-              {profile.modalities && <div>Input: {profile.modalities.input.join(", ")}</div>}
-            </div>
-          </div>
-          <div className="text-xs text-muted-foreground border-t pt-2 mt-2">
-            Profile data from{" "}
-            <a
-              href="https://models.dev"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="underline hover:text-foreground"
-            >
-              models.dev
-            </a>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function AddProviderDialog({
-  open,
-  onOpenChange,
-}: {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-}) {
-  const [name, setName] = useState("");
-  const [providerType, setProviderType] = useState<LlmProviderType>("openai");
-  const [baseUrl, setBaseUrl] = useState("");
-  const [apiKey, setApiKey] = useState("");
-
-  const createProvider = useCreateLlmProvider();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    const data: CreateLlmProviderRequest = {
-      name,
-      provider_type: providerType,
-      base_url: baseUrl || undefined,
-      api_key: apiKey || undefined,
-    };
-    await createProvider.mutateAsync(data);
-    onOpenChange(false);
-    setName("");
-    setProviderType("openai");
-    setBaseUrl("");
-    setApiKey("");
-  };
-
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Add LLM Provider</DialogTitle>
-          <DialogDescription>
-            Configure a new LLM provider for your agents to use.
-          </DialogDescription>
-        </DialogHeader>
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div className="space-y-2">
-            <Label htmlFor="name">Name</Label>
-            <Input
-              id="name"
-              value={name}
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) => setName(e.target.value)}
-              placeholder="My OpenAI Provider"
-              required
-            />
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="provider-type">Provider Type</Label>
-            <Select
-              value={providerType}
-              onValueChange={(v) => setProviderType(v as LlmProviderType)}
-            >
-              <SelectTrigger className="w-full">
-                <div className="flex items-center gap-2">
-                  <ProviderIcon providerType={providerType} size="sm" showBackground={false} />
-                  <span>{getProviderLabel(providerType)}</span>
-                </div>
-              </SelectTrigger>
-              <SelectContent>
-                {PROVIDER_TYPES.map((type) => (
-                  <SelectItem key={type.value} value={type.value}>
-                    <div className="flex items-center gap-2">
-                      <ProviderIcon providerType={type.value} size="sm" showBackground={false} />
-                      <span>{type.label}</span>
-                    </div>
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="base-url">Base URL (optional)</Label>
-            <Input
-              id="base-url"
-              value={baseUrl}
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) => setBaseUrl(e.target.value)}
-              placeholder="https://api.openai.com/v1"
-            />
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="api-key">API Key (optional)</Label>
-            <Input
-              id="api-key"
-              type="password"
-              value={apiKey}
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) => setApiKey(e.target.value)}
-              placeholder={getApiKeyPlaceholder(providerType)}
-            />
-          </div>
-          <DialogFooter>
-            <Button type="button" variant="outline" onClick={() => onOpenChange(false)}>
-              Cancel
-            </Button>
-            <Button type="submit" disabled={createProvider.isPending || !name}>
-              {createProvider.isPending ? "Creating..." : "Create Provider"}
-            </Button>
-          </DialogFooter>
-        </form>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-function SetApiKeyDialog({
-  provider,
-  open,
-  onOpenChange,
-}: {
-  provider: LlmProvider | null;
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-}) {
-  const [apiKey, setApiKey] = useState("");
-  const updateProvider = useUpdateLlmProvider(provider?.id || "");
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!provider) return;
-    await updateProvider.mutateAsync({ api_key: apiKey });
-    onOpenChange(false);
-    setApiKey("");
-  };
-
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>{provider?.api_key_set ? "Update" : "Set"} API Key</DialogTitle>
-          <DialogDescription>
-            {provider?.api_key_set
-              ? "Enter a new API key to replace the existing one."
-              : "Enter the API key for this provider."}
-          </DialogDescription>
-        </DialogHeader>
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div className="space-y-2">
-            <Label htmlFor="new-api-key">API Key</Label>
-            <Input
-              id="new-api-key"
-              type="password"
-              value={apiKey}
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) => setApiKey(e.target.value)}
-              placeholder={provider ? getApiKeyPlaceholder(provider.provider_type) : "your-api-key"}
-              required
-            />
-          </div>
-          <DialogFooter>
-            <Button type="button" variant="outline" onClick={() => onOpenChange(false)}>
-              Cancel
-            </Button>
-            <Button type="submit" disabled={updateProvider.isPending || !apiKey}>
-              {updateProvider.isPending ? "Saving..." : "Save API Key"}
-            </Button>
-          </DialogFooter>
-        </form>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-function AddModelDialog({
-  providers,
-  open,
-  onOpenChange,
-}: {
-  providers: LlmProvider[];
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-}) {
-  const [providerId, setProviderId] = useState("");
-  const [modelId, setModelId] = useState("");
-  const [displayName, setDisplayName] = useState("");
-  const [installed, setInstalled] = useState(true);
-
-  const createModel = useCreateLlmModel(providerId);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    const data: CreateLlmModelRequest = {
-      model_id: modelId,
-      display_name: displayName,
-      installed,
-    };
-    await createModel.mutateAsync(data);
-    onOpenChange(false);
-    setProviderId("");
-    setModelId("");
-    setDisplayName("");
-    setInstalled(true);
-  };
-
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Add Model</DialogTitle>
-          <DialogDescription>Add a new model to an existing provider.</DialogDescription>
-        </DialogHeader>
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div className="space-y-2">
-            <Label htmlFor="provider">Provider</Label>
-            <Select value={providerId} onValueChange={setProviderId}>
-              <SelectTrigger className="w-full">
-                <span className={!providerId ? "text-muted-foreground" : ""}>
-                  {providerId
-                    ? providers.find((p) => p.id === providerId)?.name
-                    : "Select provider"}
-                </span>
-              </SelectTrigger>
-              <SelectContent>
-                {providers.map((p) => (
-                  <SelectItem key={p.id} value={p.id}>
-                    {p.name}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="model-id">Model ID</Label>
-            <Input
-              id="model-id"
-              value={modelId}
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) => setModelId(e.target.value)}
-              placeholder="gpt-5.2"
-              required
-            />
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="display-name">Display Name</Label>
-            <Input
-              id="display-name"
-              value={displayName}
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) => setDisplayName(e.target.value)}
-              placeholder="GPT-4o"
-              required
-            />
-          </div>
-          <div className="flex items-center gap-2">
-            <input
-              type="checkbox"
-              id="model-installed"
-              checked={installed}
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) => setInstalled(e.target.checked)}
-              className="h-4 w-4"
-            />
-            <Label htmlFor="model-installed">Install model (available in UI model pickers)</Label>
-          </div>
-          <DialogFooter>
-            <Button type="button" variant="outline" onClick={() => onOpenChange(false)}>
-              Cancel
-            </Button>
-            <Button
-              type="submit"
-              disabled={createModel.isPending || !providerId || !modelId || !displayName}
-            >
-              {createModel.isPending ? "Creating..." : "Create Model"}
-            </Button>
-          </DialogFooter>
-        </form>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-function ProviderCardSkeleton() {
-  return (
-    <Card>
-      <CardHeader className="flex flex-row items-start justify-between space-y-0">
-        <div className="flex items-center gap-3">
-          <Skeleton className="h-9 w-9 rounded-lg" />
-          <div className="space-y-2">
-            <Skeleton className="h-5 w-32" />
-            <Skeleton className="h-4 w-24" />
-          </div>
-        </div>
-        <Skeleton className="h-5 w-16" />
-      </CardHeader>
-      <CardContent>
-        <Skeleton className="h-4 w-full mb-4" />
-        <Skeleton className="h-8 w-24 ml-auto" />
-      </CardContent>
-    </Card>
-  );
-}
+import { ProviderCard, ProviderCardSkeleton } from "./provider-card";
+import { ModelRow } from "./model-row";
+import { AddProviderDialog, SetApiKeyDialog, AddModelDialog } from "./provider-dialogs";
 
 export default function ProvidersPage() {
   const queryClient = useQueryClient();
