+++ apps/ui/src/lib/api/types/command-types.ts
@@ -1,0 +1,20 @@
+// Command types
+
+export type CommandSource = "system" | "skill";
+
+export interface CommandArg {
+  name: string;
+  description: string;
+  required: boolean;
+}
+
+export interface CommandDescriptor {
+  name: string;
+  description: string;
+  source: CommandSource;
+  args?: CommandArg[];
+}
+
+export interface CommandsResponse {
+  commands: CommandDescriptor[];
+}
