+++ apps/ui/src/lib/api/types/index.ts
@@ -1,0 +1,18 @@
+// Barrel re-export for all API types
+// Each domain has its own file; this index re-exports everything
+// so that `import { ... } from "@/lib/api/types"` continues to work.
+
+export * from "./common-types";
+export * from "./agent-types";
+export * from "./session-types";
+export * from "./message-types";
+export * from "./event-types";
+export * from "./auth-types";
+export * from "./llm-types";
+export * from "./capability-types";
+export * from "./durable-types";
+export * from "./mcp-types";
+export * from "./skill-types";
+export * from "./image-types";
+export * from "./app-types";
+export * from "./command-types";
